requirements.txt
tritonclient[http]==2.52.0
pillow==11.0.0
transformers==4.47.0
requests==2.32.3
huggingface_hub==0.25.2
numpy==1.26.4